compat: add interpret actions for the various pointer movement keysyms
This should automatically makes those keysyms work in Xorg (alas not in
Wayland because libxkbcommon does not handle PointerKeys).

The various button actions already had interpret actions defined.

Except... this doesn't work due to how pointerkeys has been somewhat
hacked into the X server via the keypad. Because our default compat
includes interpret actions for e.g. KP_7/KP_Up and they always resolve
to MovePtr(), the X server has a _FixupAction hack to hack any pointer
action to XkbSA_NoAction when the MouseKeys control is not set.
So by default, the action resolves to the keysym but Shift+NumLock sets
that control and voila, it all magically works.

This of course does not work happily if we don't use the MouseKeys
control (because, well, we have explicit keysyms for this now, right?).

Removing the _FixupAction gives us the desired pointer control but that
will break the old approach. Sigh, can't win...