compat: split the hardcoded numpad mousekeys into a separate compat section
This removes the interpret actions for the various mouse keys from the
default compat section so our keypad by default is just a keypad.

With option "keypad:pointerkeys" the new section is loaded which will
override the keypad with all the old actions.

The only reason this worked before was because the server checks whether
the MouseKeys control is enabled and, if not, resets to XkbSA_NoAction
for any XKB pointer action.

This should've really been done as part of
1bc01a5eaa1180c026ebf8fc75e9bc3cb6876c66