symbols/es: Better placement for dead_caron/doubleacute
Move dead_caron to the same key as dead_circumflex in Spanish layout,
since a caron is visually similar to a circumflex, but upside down.

Move dead_doubleacute to the same key as dead_acute and dead_diaeresis,
since a double acute is visually similar to an acute and a diaeresis
(and semantically related to both, as it acts as an "acute umlaut").

Move dead_abovering to where dead_doubleacute was (Ntilde key).
(This doesn't have any pros or cons over its current placement.)