Add Chinese Bopomofo (Zhuyin) layout
This is the de-facto standard keyboard layout of Taiwan.
The layout in this commit is Apple's variant,
as shown here:
https://store.storeimages.cdn-apple.com/4982/as-images.apple.com/is/MXCL3TA?wid=1144&hei=1144&fmt=jpeg&qlt=90&.v=1729134348404
(Microsoft's version lacks many Chinese punctuation characters.)

Typing Chinese characters with this layout requires a Zhuyin IME.
Traditionally, the way these have been implemented on Linux (and Windows)
is to layer directly on top of US QWERTY, which is why
this layout was never added to `xkeyboard-config`.
Mac OS has a different approach, where the layout produces Bopomofo
directly and the input method then converts that into Chinese ideograms.
This commit is therefore mostly an experiment to see what the Mac approach
would look like in XKB. (Though, even without a supporting Zhuyin IME,
this layout is still useful for inputting Bopomofo directly.)

Signed-off-by: Jules Bertholet <julesbertholet@quoi.xyz>