Implement F19 and F24
This xkeyboard-config patch is complementary to this kernel patch:
https://lore.kernel.org/all/20250311180643.1107430-2-wse@tuxedocomputers.com/

This patch implements the currently unmapped F19 and F24 keys.

F19 is mapped straight forward.

F24 is mapped to F24 or, in combination with mod keys, to touchpad toggle as
some notebooks send Control + Super + F24 upon pressing touchpad toggle. Before
the kernel patch above this was incorrectly mapped to Control + Super + <ZEHA>.
The Control + Super + <ZEHA> mapping is not removed however to stay compatible
with old kernels.

Signed-off-by: Werner Sembach <wse@tuxedocomputers.com>