options: Added hyper:mod2 compatibility option
This is a compatibility option for enabling `Hyper`, in addition to
`hyper:mod3` and `hyper:mod4`.

`hyper:mod2` maps the virtual modifier `Hyper` to `Mod2` and then
*conflicts with `NumLock`*.

Use this option if using `Hyper`, `Super` *and* `LevelFive`, e.g. for
layouts with 5+ levels. The option `numpad:mac` should be activated as
well, in order to enable the numpad.