Fix shift:break_caps resulting in rEVERSE cAPS
It happens if some keys were operated simultaneously with the `Shift`
keys when trying to unlock `Caps`.

However, the usual workflow expected with this option is:

- `<CapsLock>` ↓↑: `Caps` is locked
- (type caps letters)
- `<Shift>` ↓: `Caps` is unlocked, `Shift` is set
- (optionally type some more upper case letters: tolerance to sync with
  `Shift`)
- `<Shift>` ↑: `Shift` is unset
- type in lower case

The root issue is that in the current code:

```c
xkb_symbols "breaks_caps" {
    key <LFSH> {
	type[group1] = "ALPHABETIC",
	symbols[group1] = [ Shift_L, Shift_L ],
	actions[group1] = [ SetMods(modifiers=Shift),
			    SetMods(modifiers=Shift+Lock,clearLocks) ]
    };
    // […]
};
```

the `clearLocks` parameter triggers the “clear locks” feature *only* if
*no* other key is operated *simultaneously* with the `<Shift>` keys,
which may not happen with the key sequence mentioned above. But even if
it could trigger with other keys operated simultaneously, it would still
set *both* `Shift` and `Lock`, which make `Shift` cancel `Lock` on usual
key types and then results in typing lower case letters when `<Shift>`
is pressed.

Using simply `[Shift_L, Caps_Lock]` fixes the issue. In fact we already
have the corresponding option: `shift:both_capslock_cancel`.

Note that a side-effect is that pressing both Shift keys now triggers
CapsLock.