build: Do not copy files manually to build directory
Before this commit XKB files were copied “manually” to the build
directory in order to use it as a proper XKB root in tests and latin
layouts listing. But this is hacky, as Meson does not track these
copied files.

Instead use a *list* of XKB roots, populated in order with:
1. Build directory: takes precedence over the source directory.
   It contains generated files such as rules, registries and symbols.
   `XKB_CONFIG_EXTRA_PATH` = `meson.project_build_root()`
2. Source directory: used as a fallback for any non-generated
   file/section (most of the KcCGST files).
   `XKB_CONFIG_ROOT` = `meson.project_source_root()`;

These values are read by the tests.

Part-of: <https://gitlab.freedesktop.org/xkeyboard-config/xkeyboard-config/-/merge_requests/858>